fix: address reserve yield review feedback
